Refactored Transformation interfaces into their own package to de-clutter the top level and fixed some generics using Raw types
