+++ src/main/java/uk/co/edstow/cain/goals/atomGoal/pairGen/SimplePairGenFactory.java
@@ -7,7 +7,7 @@ import uk.co.edstow.cain.pairgen.PairGenFactory;
 import uk.co.edstow.cain.structures.GoalBag;
 import uk.co.edstow.cain.structures.GoalPair;
 import uk.co.edstow.cain.util.Tuple;
-import uk.co.edstow.cain.Transformation;
+import uk.co.edstow.cain.transformations.Transformation;
 
 import java.util.*;
