+++ src/main/java/uk/co/edstow/cain/goals/atomGoal/pairGen/V2PairGenFactory.java
@@ -1,6 +1,5 @@
 package uk.co.edstow.cain.goals.atomGoal.pairGen;
 
-import uk.co.edstow.cain.Transformation;
 import uk.co.edstow.cain.goals.atomGoal.Atom;
 import uk.co.edstow.cain.goals.atomGoal.AtomGoal;
 import uk.co.edstow.cain.pairgen.Context;
