+++ src/main/java/uk/co/edstow/cain/nonlinear/NonLinearGoal.java
@@ -1,6 +1,6 @@
 package uk.co.edstow.cain.nonlinear;
 
-import uk.co.edstow.cain.Transformation;
+import uk.co.edstow.cain.transformations.Transformation;
 import uk.co.edstow.cain.structures.Bounds;
 import uk.co.edstow.cain.structures.Goal;
 import uk.co.edstow.cain.structures.GoalPair;
