+++ src/main/java/uk/co/edstow/cain/nonlinear/NonLinearPairGenFactory.java
@@ -1,6 +1,6 @@
 package uk.co.edstow.cain.nonlinear;
 
-import uk.co.edstow.cain.Transformation;
+import uk.co.edstow.cain.transformations.Transformation;
 import uk.co.edstow.cain.pairgen.Context;
 import uk.co.edstow.cain.pairgen.PairGenFactory;
 import uk.co.edstow.cain.structures.Goal;
@@ -11,7 +11,6 @@ import uk.co.edstow.cain.util.Tuple;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import static uk.co.edstow.cain.nonlinear.NonLinearGoal.idOfGoal;
 import static uk.co.edstow.cain.nonlinear.NonLinearGoal.idOfGoals;
