+++ src/main/java/uk/co/edstow/cain/pairgen/PairGenFactory.java
@@ -1,10 +1,9 @@
 package uk.co.edstow.cain.pairgen;
 
-import uk.co.edstow.cain.Transformation;
+import uk.co.edstow.cain.transformations.Transformation;
 import uk.co.edstow.cain.structures.Goal;
 import uk.co.edstow.cain.structures.GoalBag;
 import uk.co.edstow.cain.structures.GoalPair;
-import uk.co.edstow.cain.structures.Plan;
 import uk.co.edstow.cain.util.Tuple;
 
 import java.util.ArrayList;
