+++ src/main/java/uk/co/edstow/cain/scamp5/ThresholdScamp5ConfigGetter.java
@@ -1,6 +1,7 @@
 package uk.co.edstow.cain.scamp5;
 
-import uk.co.edstow.cain.Transformation;
+import uk.co.edstow.cain.regAlloc.RegisterAllocator;
+import uk.co.edstow.cain.transformations.Transformation;
 import uk.co.edstow.cain.pairgen.Context;
 import uk.co.edstow.cain.pairgen.CostHeuristic;
 import uk.co.edstow.cain.pairgen.PairGenFactory;
@@ -9,16 +10,16 @@ import uk.co.edstow.cain.structures.GoalBag;
 
 import java.util.List;
 
-public class ThresholdScamp5ConfigGetter<G extends Goal<G>, T extends Transformation, C extends Scamp5ConfigGetter.Scamp5Config<G, C>> implements Scamp5ConfigGetter<G,T, C> {
+public class ThresholdScamp5ConfigGetter<G extends Goal<G>, T extends Transformation<R>, R extends RegisterAllocator.Register, SELF extends Scamp5ConfigGetter.Scamp5Config<G, SELF>> implements Scamp5ConfigGetter<G,T,R, SELF> {
     private final CostHeuristic<G, T> heuristic;
-    private final C scamp5Config;
-    private final C scamp5ConfigMovOnly;
+    private final SELF scamp5Config;
+    private final SELF scamp5ConfigMovOnly;
     private final int threshold;
-    private final GenGetter<G,T,C> above;
-    private final GenGetter<G,T,C> below;
+    private final GenGetter<G,T,R, SELF> above;
+    private final GenGetter<G,T,R, SELF> below;
 
 
-    public ThresholdScamp5ConfigGetter(List<G> initialGoals, int threshold, CostHeuristic<G, T> heuristic, C scamp5Config, GenGetter<G, T, C> above, GenGetter<G, T, C> below) {
+    public ThresholdScamp5ConfigGetter(List<G> initialGoals, int threshold, CostHeuristic<G, T> heuristic, SELF scamp5Config, GenGetter<G, T, R, SELF> above, GenGetter<G, T, R, SELF> below) {
         this.heuristic = heuristic;
         this.scamp5Config = scamp5Config;
         this.scamp5ConfigMovOnly = scamp5Config.getMovOnlyVersion();
@@ -44,12 +45,12 @@ public class ThresholdScamp5ConfigGetter<G extends Goal<G>, T extends Transforma
     }
 
     @Override
-    public C getScamp5ConfigForDirectSolve(GoalBag<G> goals, Context<G, T> context) {
+    public SELF getScamp5ConfigForDirectSolve(GoalBag<G> goals, Context<G, T> context) {
         return scamp5Config;
     }
 
     @FunctionalInterface
-    public interface GenGetter<G extends Goal<G>,T extends Transformation, C extends Scamp5ConfigGetter.Scamp5Config<G, C>>{
+    public interface GenGetter<G extends Goal<G>, T extends Transformation<R>, R extends RegisterAllocator.Register, C extends Scamp5ConfigGetter.Scamp5Config<G, C>>{
         PairGenFactory.PairGen<G, T> get(GoalBag<G> goals, Context<G, T> conf, C scamp5Config, CostHeuristic<G, T> heuristic);
     }
 }
