+++ src/main/java/uk/co/edstow/cain/structures/GoalPair.java
@@ -1,6 +1,6 @@
 package uk.co.edstow.cain.structures;
 
-import uk.co.edstow.cain.Transformation;
+import uk.co.edstow.cain.transformations.Transformation;
 
 import java.util.ArrayList;
 import java.util.Collections;
